# lib/api-zod/src/generated/types/intelligenceEntryItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IntelligenceEntryItem {
  id: number;
  title: string;
  summary?: string | null;
  status: string;
  type?: string | null;
  severity?: string | null;
  createdAt: Date;
}