# lib/api-zod/src/generated/types/projectIntelligence.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectIntelligenceDna } from './projectIntelligenceDna';
import type { ProjectIntelligenceEntries } from './projectIntelligenceEntries';
import type { ProjectIntelligenceHealth } from './projectIntelligenceHealth';
import type { ProjectIntelligenceReadiness } from './projectIntelligenceReadiness';

/**
 * The single source of truth for a project. All surfaces — Master Map, Axiom Flow header, HUD, Workspace ring, and Portfolio cards — consume this object instead of computing independent realities. Composes genome DNA, health signals, readiness %, and typed entry lists.

 */
export interface ProjectIntelligence {
  projectId: number;
  projectName?: string | null;
  projectDescription?: string | null;
  projectStatus?: string | null;
  /** Genome DNA fields */
  dna: ProjectIntelligenceDna;
  /** Computed health signals — never stored, always fresh */
  health: ProjectIntelligenceHealth;
  /** Canonical readiness score — the workspace ring and portfolio cards read from here */
  readiness: ProjectIntelligenceReadiness;
  /** Typed entry lists — Master Map, Ledger, and HUD read from here */
  entries: ProjectIntelligenceEntries;
  computedAt: Date;
}